sclang: move primregistry helper macros to separate header
