+++ lang/LangPrimSource/SC_PrimRegistry.hpp
@@ -14,6 +14,7 @@
 #include <boost/preprocessor/cat.hpp>
 
 #include "PyrPrimitive.h"
+#include "SC_PrimRegistryMacros.hpp"
 
 class SC_PrimRegistry {
 public:
@@ -76,50 +77,6 @@ struct SC_PrimRegistryHelper {
         reg.add_entry( {primitive, symbol, numArgs, varArgs} );
     }
 };
-
-// utilities
-#define SCLANG_NUMARGS_MSG "sclang primitive must have at least 1 argument"
-#define SCLANG_VARARGS_MSG "sclang primitive's varargs quality must be either 0 or 1"
-#define SCLANG_PRIMITIVE_STATIC_ASSERTS( numArgs, varArgs )                                           \
-    static_assert( numArgs > 0, SCLANG_NUMARGS_MSG );                                                 \
-    static_assert( varArgs == 0 || varArgs == 1, SCLANG_VARARGS_MSG )
-
-/// Builds the C++ primitive function name
-#define SCLANG_PRIMITIVE_NAME( name )                                                                 \
-    BOOST_PP_CAT( pr, name )
-
-/// Function signature for the C++ primitive
-#define SCLANG_PRIMITIVE_SIGNATURE( name )                                                            \
-    int SCLANG_PRIMITIVE_NAME( name ) ( struct VMGlobals *g, int numArgsPushed )
-
-/// Function signature for a templated C++ primitive
-#define SCLANG_TEMPLATE_PRIMITIVE_SIGNATURE( name, T )                                                \
-    template <typename T>                                                                             \
-    SCLANG_PRIMITIVE_SIGNATURE( name )
-
-/// SuperCollider primitive symbol
-#define SCLANG_PRIMITIVE_SYMBOL( name )                                                               \
-    BOOST_PP_STRINGIZE( BOOST_PP_CAT( _, name) )
-
-#define SCLANG_REGISTRY_HELPER_DECL( name )                                                           \
-    static SC_PrimRegistryHelper BOOST_PP_CAT( name, _RegistryHelper )
-
-#define SCLANG_REGISTRY_HELPER_DEFN( name, numArgs, varArgs )                                         \
-    SCLANG_PRIMITIVE_STATIC_ASSERTS( numArgs, varArgs );                                              \
-    SCLANG_REGISTRY_HELPER_DECL( name ) {                                                             \
-        & SCLANG_PRIMITIVE_NAME( name ),                                                              \
-        SCLANG_PRIMITIVE_SYMBOL( name ),                                                              \
-        numArgs,                                                                                      \
-        varArgs,                                                                                      \
-        SC_PrimRegistry::instance()                                                                   \
-    };
-
-/// Generates C++ function signature and registers with SC_PrimRegistry; private, do not use directly
-#define SCLANG_DEFINE_PRIMITIVE_HELPER( name, numArgs, varArgs )                                      \
-    SCLANG_PRIMITIVE_SIGNATURE( name );                                                               \
-    SCLANG_REGISTRY_HELPER_DEFN( name, numArgs, varArgs );                                            \
-    SCLANG_PRIMITIVE_SIGNATURE( name )
-
 /// Use to define an ordinary (non-varArgs) primitive.
 #define SCLANG_DEFINE_PRIMITIVE( name, numArgs )                                                      \
     SCLANG_DEFINE_PRIMITIVE_HELPER( name, numArgs, 0 )
@@ -128,32 +85,11 @@ struct SC_PrimRegistryHelper {
 #define SCLANG_DEFINE_VARARGS_PRIMITIVE( name, numArgs )                                              \
     SCLANG_DEFINE_PRIMITIVE_HELPER( name, numArgs, 1 )
 
-// ----------- template primitives -------------
-
-/// Helper, do not use directly.
-#define SCLANG_REGISTER_TEMPLATE_PRIMITIVE( name, numArgs, varArgs, Type )                            \
-    SCLANG_PRIMITIVE_STATIC_ASSERTS( numArgs, varArgs );                                              \
-    SCLANG_REGISTRY_HELPER_DECL( BOOST_PP_CAT( name, BOOST_PP_CAT(_, Type) ) ) {                      \
-        & SCLANG_PRIMITIVE_NAME( name )<Type>,                                                        \
-        SCLANG_PRIMITIVE_SYMBOL( BOOST_PP_CAT( name, BOOST_PP_CAT(_, Type) ) ),                       \
-        numArgs,                                                                                      \
-        varArgs,                                                                                      \
-        SC_PrimRegistry::instance()                                                                   \
-    }
-
-/// Helper, do not use directly.
-#define SCLANG_REGISTER_TEMPLATE_PRIMITIVE_HELPER( r, data, elem )                                    \
-    SCLANG_REGISTER_TEMPLATE_PRIMITIVE(                                                               \
-        BOOST_PP_SEQ_ELEM(0, data),                                                                   \
-        BOOST_PP_SEQ_ELEM(1, data),                                                                   \
-        BOOST_PP_SEQ_ELEM(2, data),                                                                   \
-        elem);
-
 /**
  * Use to define a templated primitive.
  *
- * T is the typename for the template, and TypeSeq is a preprocessor-seq of type names with which
- * the primitive will be instantiated. TypeSeq is of the form "(T1)(T2)(T3)" (no quotes).
+ * \c T is the typename for the template, and \c TypeSeq is a preprocessor-seq of type names with
+ * which the primitive will be instantiated. TypeSeq is of the form "(T1)(T2)(T3)" (no quotes).
  */
 #define SCLANG_DEFINE_TEMPLATE_PRIMITIVE( name, T, numArgs, TypeSeq )                                 \
     SCLANG_TEMPLATE_PRIMITIVE_SIGNATURE( name, T );                                                   \
@@ -161,13 +97,6 @@ struct SC_PrimRegistryHelper {
     SCLANG_TEMPLATE_PRIMITIVE_SIGNATURE( name, T )
 
 // ------------ libsclang macros -------------
-
-// utility
-#define LIBSCLANG_DUMMY_VAR( name )                                                                   \
-    libsclang_dummy_var_ ## name
-
-#define LIBSCLANG_DUMMY_VAR_LOCAL( name )                                                             \
-    libsclang_dummy_var_ ## name ## _local
 
 /// Use at the top of a file containing primitives
 #define LIBSCLANG_PRIMITIVE_GROUP( name )                                                             \
